requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.12
openai==1.12.0
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
gunicorn==21.2.0 